pyproject.toml
[tool.poetry]
name = "hotel-api"
version = "0.1.0"
description = "Simple hotel rooms & bookings API (Django + DRF)"
authors = ["You <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
Django = "^5.0"
djangorestframework = "^3.14"
psycopg2-binary = "^2.9"
gunicorn = "^21.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-django = "^4.8"
pytest-cov = "^5.0"
requests = "^2.32"
ruff = "^0.6"
black = "^24.0"
pre-commit = "^3.7"

[build-system]
requires = ["poetry-core>=1.7.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.black]
line-length = 88
target-version = ["py311"]
